+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # jojo changelog
+## 1.0.0
+- Switched order of `rawDate` and `date`
+- `config.dataParser`, `config.formatter` no longer `require` based
+
 0.2.0 - Added test suite and Travis CI. Scrapped old folder structure.
 
 Before 0.1.7 - See `git log`